src/lab/requirements.txt
pydantic==1.10.2
pillow==9.3.0
fastapi==0.87.0
uvicorn[standard]==0.20.0
opencv-python==4.6.0.66
